# docs/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Warehouse Management System API",
    "description": "API для системы управления складами, товарами и инвентаризацией с аналитикой продаж",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/api",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
...
  "paths": {
...
    "/warehouses": {
      "get": {
        "tags": ["warehouses"],
        "summary": "Получить список всех складов",
        "description": "Возвращает список всех складов в системе",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Список складов",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Warehouse"
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["warehouses"],
        "summary": "Создать новый склад",
        "description": "Создает новый склад в системе",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "warehouse",
            "in": "body",
            "description": "Информация о складе",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Warehouse"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Склад успешно создан",
            "schema": {
              "$ref": "#/definitions/Warehouse"
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/products": {
      "get": {
        "tags": ["products"],
        "summary": "Получить список всех товаров",
        "description": "Возвращает список всех товаров в системе",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Список товаров",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["products"],
        "summary": "Создать новый товар",
        "description": "Создает новый товар в системе",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "product",
            "in": "body",
            "description": "Информация о товаре",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Товар успешно создан",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "put": {
        "tags": ["products"],
        "summary": "Обновить товар",
        "description": "Обновляет информацию о существующем товаре",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID товара",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "product",
            "in": "body",
            "description": "Обновленная информация о товаре",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Товар успешно обновлен",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/inventory": {
      "post": {
        "tags": ["inventory"],
        "summary": "Создать запись инвентаризации",
        "description": "Добавляет товар на склад",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "inventory",
            "in": "body",
            "description": "Информация о товаре на складе",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Inventory"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Запись инвентаризации успешно создана",
            "schema": {
              "$ref": "#/definitions/Inventory"
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/quantity": {
      "put": {
        "tags": ["inventory"],
        "summary": "Обновить количество товара на складе",
        "description": "Изменяет количество определенного товара на указанном складе",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "updateRequest",
            "in": "body",
            "description": "Информация для обновления количества",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "warehouse_id": {
                  "type": "string",
                  "format": "uuid",
                  "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "product_id": {
                  "type": "string",
                  "format": "uuid",
                  "example": "123e4567-e89b-12d3-a456-426614174001"
                },
                "quantity": {
                  "type": "integer",
                  "example": 5
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Количество товара успешно обновлено",
            "schema": {
              "$ref": "#/definitions/Inventory"
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/inventory/discount": {
      "put": {
        "tags": ["inventory"],
        "summary": "Обновить скидку на товар",
        "description": "Устанавливает скидку на указанный товар на складе",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "updateRequest",
            "in": "body",
            "description": "Информация для обновления скидки",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "warehouse_id": {
                  "type": "string",
                  "format": "uuid",
                  "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "product_id": {
                  "type": "string",
                  "format": "uuid",
                  "example": "123e4567-e89b-12d3-a456-426614174001"
                },
                "discount": {
                  "type": "number",
                  "format": "float",
                  "example": 15
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Скидка успешно обновлена",
            "schema": {
              "$ref": "#/definitions/Inventory"
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/warehouses/{id}/products": {
      "get": {
        "tags": ["warehouses", "inventory"],
        "summary": "Получить список товаров на складе",
        "description": "Возвращает список товаров на указанном складе с поддержкой пагинации",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID склада",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Номер страницы (начиная с 1)",
            "required": false,
            "type": "integer",
            "default": 1
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Количество записей на странице",
            "required": false,
            "type": "integer",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Список товаров на складе",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/InventoryWithProduct"
              }
            }
          },
          "400": {
            "description": "Некорректный запрос",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Внутренняя ошибка сервера",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Warehouse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid",
          "example": "123e4567-e89b-12d3-a456-426614174000"
        },
        "address": {
          "type": "string",
          "example": "ул. Складская, 123"
        }
      }
    },
    "Product": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid",
          "example": "123e4567-e89b-12d3-a456-426614174001"
        },
        "name": {
          "type": "string",
          "example": "Ноутбук"
        },
        "description": {
          "type": "string",
          "example": "Ноутбук Dell XPS 13"
        },
        "characteristics": {
          "type": "object",
          "example": {
            "processor": "Intel i7",
            "ram": "16GB",
            "storage": "512GB SSD"
          }
        },
        "weight": {
          "type": "number",
          "format": "float",
          "example": 1.3
        },
        "barcode": {
          "type": "string",
          "example": "1234567890123"
        }
      }
    },
...
  }
}